% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayDataTable.R
\name{displayDataTable}
\alias{displayDataTable}
\title{function to add delete button to the datatable}
\usage{
displayDataTable(df, id1, id2, ...)
}
\arguments{
\item{df}{data frame}

\item{id}{id prefix to add to each actionButton. The buttons will be id'd as id_INDEX.}
}
\value{
A DT::datatable with escaping turned off that has the delete buttons in the first column and \code{df} in the other
}
\description{
A column of delete buttons for each row in the data frame for the first column
}
\keyword{internal}
