% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBrightness.R
\name{getBrightness}
\alias{getBrightness}
\title{The brightness map of an image (0 to 1).}
\usage{
getBrightness(rgbArray)
}
\arguments{
\item{rgbArray}{a 3-column array of R, G abd B color channels}
}
\value{
brightness matrix
}
\description{
The brightness map of an image (0 to 1).
}
\keyword{internal}
