% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExampleData.R
\name{getExampleData}
\alias{getExampleData}
\title{Downloads and installs the WIAD Example Dataset}
\usage{
getExampleData(
  mode = "install",
  downloadPath = tempdir(),
  installPath = "",
  dataset = "small"
)
}
\arguments{
\item{mode}{\code{"install"} (default),\code{"update"} or \code{"remove"}}

\item{downloadPath}{custom location to temporarily store WIAD example dataset while downloading. Default: \code{tempdir()}}

\item{installPath}{custom location to place WIAD example dataset. Default: WIAD package directory}

\item{dataset}{choose to download the \code{"small"} example dataset (248 MB) or the  \code{"full"} example dataset (1.83 GB). Default: \code{"small"}}
}
\value{
No return value
}
\description{
The WIAD Example Dataset can be downloaded from https://wiad.science/WIAD_ExampleData.zip .
This function downloads and extracts the WIAD Example Dataset in the package folder.
}
