% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWIADExampleData.R
\name{getWIADExampleData}
\alias{getWIADExampleData}
\title{Downloads and installs the WIAD Example Dataset}
\usage{
getWIADExampleData(mode = "install", downloadPath = "", installPath = "")
}
\arguments{
\item{mode}{\code{"install"} (default),\code{"update"} or \code{"remove"}}

\item{downloadPath}{custom location to temporarily store WIAD example dataset while downloading. Default: tempdir()}

\item{installPath}{custom location to place WIAD example dataset. Default: WIAD package directory}
}
\description{
The WIAD Example Dataset can be downloaded from http://wiad.science/WIAD_ExampleData.zip .
This function downloads and extracts the WIAD Example Dataset in the package folder.
}
