% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Launch.R
\name{Launch}
\alias{Launch}
\title{Launch WIAD interface}
\usage{
Launch(archiveDir = "./WIAD_ARCHIVE/", Interactive = FALSE)
}
\arguments{
\item{archiveDir}{path to the archive directory}

\item{Interactive}{logical variable to force an interactive session}
}
\value{
No return value, this should be run in an interactive R session
}
\description{
This function launches the app by opening the default web browser.
}
\examples{

#Launch WIAD interface
wiad::Launch()


}
