% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRowNumber.R
\name{parseRowNumber}
\alias{parseRowNumber}
\title{Extracts the row id number from the id string to delete individual rows}
\usage{
parseRowNumber(idstr)
}
\arguments{
\item{idstr}{the id string formated as id_INDEX}
}
\value{
INDEX from the id string id_INDEX
}
\description{
Extracts the row id number from the id string to delete individual rows
}
\keyword{internal}
