% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotateRGB.R
\name{rotateRGB}
\alias{rotateRGB}
\title{rotate an RGB array (or any 3D array)}
\usage{
rotateRGB(imgMat)
}
\arguments{
\item{imgMat}{input RGB array}
}
\value{
rotated each color channel
}
\description{
rotate an RGB array (or any 3D array)
}
\keyword{internal}
